gh-118908: Use __main__ for the default PyREPL namespace
This allows for PYTHONSTARTUP to execute in the right namespace.
